requirements.app.txt
pandas>=2.0.0
psycopg2-binary>=2.9.0
boto3==1.37.38
kafka-python==2.1.2
marimo==0.23.1

